Add handler sampling functions
They provide abstraction for sampling, regardless of storage engine.
Right now, sampling can only be done if the table has a Primary Key
and the storage engine must be InnoDB. The indexes are sorted and
the sampling positions are chosen between the min and max PK values.
Sampling is deterministic and the values span the whole range
of the PK based on its min and max values (see
handler::ha_rnd_sample function for details).